patch: begin fixtures.
This patching system will be based heavily on JSON Patch (RFC 6902).
Many examples will come from that RFC.

Some divergences will also occur.  For example, these fixtures already
include some assertions that map order will be stable, which was not
maintained in RFC 6902.  It is also likely that we will *not* use the
path escaping system specified in RFC 6902 (and moreso in RFC 6901);
to use tildes for escaping seems *very* strange.

An implementation is developed in go-ipld-prime in parallel with
the growth of these fixtures.  Other implementations should be able
to follow suit in the future.